Move bare-metal common support code to dedicated files
This allows re-using the code for other bare-metal programs.
