+++ bare-metal.lds
